Refactor de SystemTestCase
